requirements.txt
absl-py==2.1.0
astor==0.4.0
gast==0.5.4
grpcio==1.62.0
joblib==0.13.2
Keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
keras-squeezenet==0.4  # Resolved conflict
Markdown==3.4.4
mock==3.0.5
numpy==1.21.6
opencv-python==4.9.0.80
protobuf==3.19.6
PyYAML==6.0.1
scikit-learn==0.21.2
scipy==1.7.3
six==1.16.0
tensorboard==2.11.12  # Resolved conflict
tensorflow==2.11.0
tensorflow-estimator==2.11.0
termcolor==2.3.0
Werkzeug==0.15.4
setuptools~=69.1.1  # Resolved conflict
Flask~=2.2.5
Flask-SQLAlchemy~=3.0.5
SQLAlchemy~=2.0.28
requests~=2.31.0
mysql-connector-python~=8.0.33
PyMySQL~=1.1.0